Extract monitor to its own module
