+++ test/generic_cache_test.exs
@@ -377,17 +377,17 @@ defmodule GenericCacheTest do
 
     {:ok, "V1"} = PDLTestGenericCache.get(1)
 
-    previous_pid = Process.whereis(pdl.cache_state_manager)
+    previous_pid = Process.whereis(pdl.cache_monitor)
 
-    :ok = GenServer.stop(pdl.cache_state_manager)
+    :ok = GenServer.stop(pdl.cache_monitor)
 
     # make sure that it restarts via the supervisor
     wait_for(fn ->
-      pid = Process.whereis(pdl.cache_state_manager)
+      pid = Process.whereis(pdl.cache_monitor)
       is_pid(pid) && Process.alive?(pid)
     end)
 
-    refute previous_pid == Process.whereis(pdl.cache_state_manager)
+    refute previous_pid == Process.whereis(pdl.cache_monitor)
 
     # cache state is unchanged
     assert {:ok, "V1"} == PDLTestGenericCache.get(1)
